weiqi-test/weiqi-sgf/requirements.txt
pytest-playwright>=0.4.0
pytest-html>=4.1.0
